No need for a separate version of CC Stereo anymore: both types of OpenGL widgets can now be managed in parallel
