+++ CMakeLists.txt
@@ -20,9 +20,6 @@ include( DeployQt )
 # CCViewer
 option( OPTION_BUILD_CCVIEWER "Check to compile CCViewer project" ON )
 
-# Quad buffer stereo support
-option( OPTION_GL_QUAD_BUFFER_SUPPORT "Check to compile CloudCompare and ccViewer with Quad Buffer support" OFF )
-
 # Testing
 option( BUILD_TESTING "Build tests for CC" OFF )
 if ( BUILD_TESTING )
@@ -38,10 +35,6 @@ set( CMAKE_DEBUG_POSTFIX "d" )
 if(WIN32 OR APPLE) 
 	set( CLOUDCOMPARE_DEST_FOLDER CloudCompare )
 	set( CCVIEWER_DEST_FOLDER ccViewer )
-	if (OPTION_GL_QUAD_BUFFER_SUPPORT)
-		set ( CLOUDCOMPARE_DEST_FOLDER ${CLOUDCOMPARE_DEST_FOLDER}Stereo )
-		set ( CCVIEWER_DEST_FOLDER ${CCVIEWER_DEST_FOLDER}Stereo )
-	endif()
 else()
 	set( CLOUDCOMPARE_DEST_FOLDER bin )
 	set( CCVIEWER_DEST_FOLDER bin )
