+++ libs/qCC_glWindow/CMakeLists.txt
@@ -7,10 +7,6 @@ add_subdirectory( src )
 
 target_compile_definitions( ${PROJECT_NAME} PRIVATE CCGLWINDOW_LIB_BUILD )
 
-if ( OPTION_GL_QUAD_BUFFER_SUPPORT )
-	target_compile_definitions( ${PROJECT_NAME} PRIVATE CC_GL_WINDOW_USE_QWINDOW )
-endif()
-
 target_link_libraries( ${PROJECT_NAME}
 	QCC_DB_LIB
 )
