+++ libs/qCC_glWindow/src/CMakeLists.txt
@@ -4,6 +4,7 @@ target_sources( ${PROJECT_NAME}
 	    ${CMAKE_CURRENT_LIST_DIR}/ccRenderingTools.cpp
 		${CMAKE_CURRENT_LIST_DIR}/ccGLWindow.cpp
 		${CMAKE_CURRENT_LIST_DIR}/ccGLWindowInterface.cpp
+		${CMAKE_CURRENT_LIST_DIR}/ccGLWindowStereo.cpp
 		${CMAKE_CURRENT_LIST_DIR}/ccGuiParameters.cpp
 		${CMAKE_CURRENT_LIST_DIR}/ccGLUtils.cpp
 )
